+++ src/relstorage/adapters/connections.py
@@ -84,6 +84,17 @@ class AbstractManagedConnection(object):
             self.active = True
         return self._cursor
 
+    def enter_critical_phase_until_transaction_end(self):
+        """
+        Given an already opened connection, cause it to attempt to
+        raise it's priority and return results faster.
+
+        This mostly has meaning for gevent drivers.
+        """
+        self.connmanager.driver.enter_critical_phase_until_transaction_end(
+            *self.open_if_needed()
+        )
+
     def drop(self):
         self.active = False
         conn, cursor = self.connection, self._cursor
@@ -297,3 +308,4 @@ class ClosedConnection(object):
         raise NotImplementedError
 
     restart_and_call = isolated_connection
+    enter_critical_phase_until_transaction_end = isolated_connection