+++ src/relstorage/cache/local_client.py
@@ -505,22 +505,30 @@ class LocalClient(object):
         entry = self._peek(oid)
         return entry is not None and entry % tid is not None
 
-    def get(self, oid_tid, update_mru=True):
+    def get(self, oid_tid, peek=False):
         oid, tid = oid_tid
         assert tid is None or tid >= 0
         decompress = self._decompress
         value = None
 
-        with self._lock:
+        if peek:
+            # Avoid the lock. We assume the underlying
+            # operations (dict.__getitem__, entry.__mod__)
+            # are thread-safe, when compared to other operations
+            # that might mutate.
             entry = self._peek(oid)
             if entry is not None:
                 value = entry % tid
-            if value is not None:
-                self._hits += 1
-                if update_mru:
+        else:
+            with self._lock:
+                entry = self._peek(oid)
+                if entry is not None:
+                    value = entry % tid
+                if value is not None:
+                    self._hits += 1
                     self._cache_mru(oid) # Make an actual hit.
-            else:
-                self._misses += 1
+                else:
+                    self._misses += 1
 
         # Finally, while not holding the lock, decompress if needed.
         # Recall that for deleted objects, `state` can be None.